Passing website for desktop during copilot panel initialization.
