+++ src/client/power-pages/actions-hub/CurrentSiteContext.ts
@@ -4,8 +4,9 @@
  */
 
 import * as vscode from 'vscode';
-import { getCurrentSiteInfo } from '../../../common/utilities/Utils';
 import { ICurrentSiteContext } from '../../../common/constants';
+import path from 'path';
+import { findWebsiteYmlFolder, getWebsiteRecordId } from '../../../common/utilities/WorkspaceInfoFinderUtil';
 
 class CurrentSiteContext implements ICurrentSiteContext {
     private _currentSiteId: string | null = null;
@@ -13,22 +14,33 @@ class CurrentSiteContext implements ICurrentSiteContext {
     private _disposables: vscode.Disposable[] = [];
     private readonly _onChanged = new vscode.EventEmitter<ICurrentSiteContext>();
     public readonly onChanged = this._onChanged.event;
+    private _isInitialized = false;
 
     constructor() {
-        const { currentSiteId, currentSiteFolderPath } = getCurrentSiteInfo();
-        this._currentSiteId = currentSiteId;
-        this._currentSiteFolderPath = currentSiteFolderPath;
-
+        // Initialize with default values, then update asynchronously
+        this._initialize();
+        
         this._disposables.push(
             vscode.window.onDidChangeActiveTextEditor(() => {
-                const newSiteInfo = getCurrentSiteInfo();
-                if (newSiteInfo.currentSiteId !== this._currentSiteId) {
-                    this._currentSiteId = newSiteInfo.currentSiteId;
-                    this._currentSiteFolderPath = newSiteInfo.currentSiteFolderPath;
-                    this._onChanged.fire(newSiteInfo);
-                }
+                this._updateSiteContext();
             })
         );
+    }
+
+    private async _initialize(): Promise<void> {
+        const siteInfo = getCurrentSiteInfo();
+        this._currentSiteId = siteInfo.currentSiteId;
+        this._currentSiteFolderPath = siteInfo.currentSiteFolderPath;
+        this._isInitialized = true;
+    }
+
+    private async _updateSiteContext(): Promise<void> {
+        const newSiteInfo = getCurrentSiteInfo();
+        if (newSiteInfo.currentSiteId !== this._currentSiteId) {
+            this._currentSiteId = newSiteInfo.currentSiteId;
+            this._currentSiteFolderPath = newSiteInfo.currentSiteFolderPath;
+            this._onChanged.fire(newSiteInfo);
+        }
     }
 
     public get currentSiteId(): string | null {
@@ -39,10 +51,39 @@ class CurrentSiteContext implements ICurrentSiteContext {
         return this._currentSiteFolderPath;
     }
 
-    public dispose() {
+    public get isInitialized(): boolean {
+        return this._isInitialized;
+    }
+
+    public dispose(): void {
         this._onChanged.dispose();
         this._disposables.forEach(d => d.dispose());
     }
 }
 
 export default new CurrentSiteContext();
+
+export function getCurrentSiteInfo():ICurrentSiteContext {
+    const activeEditor = vscode.window.activeTextEditor;
+    if (activeEditor) {
+        const filePath = activeEditor.document.uri.fsPath;
+        const fileDirectory = path.dirname(filePath);
+        const websiteYmlFolder = findWebsiteYmlFolder(fileDirectory);
+
+        if (websiteYmlFolder) {
+            const siteId = getWebsiteRecordId(websiteYmlFolder);
+            if (siteId) {
+                return { currentSiteId: siteId, currentSiteFolderPath: websiteYmlFolder };
+            }
+        }
+    }
+
+    // Fallback: check first workspace folder
+    const workspaceFolders = vscode.workspace.workspaceFolders;
+    if (workspaceFolders && workspaceFolders.length > 0) {
+        const siteId = getWebsiteRecordId(workspaceFolders[0].uri.fsPath);
+        return { currentSiteId: siteId, currentSiteFolderPath: workspaceFolders[0].uri.fsPath };
+    }
+
+    return { currentSiteId: null, currentSiteFolderPath: null };
+}
